+++ spec/integration/rendering_spec.rb
@@ -5,6 +5,7 @@ RSpec.describe "Rendering" do
     expected = <<~DESC
       Commands:
         foo assets [SUBCOMMAND]
+        foo callbacks DIR                      # Command with callbacks
         foo console                            # Starts Foo console
         foo db [SUBCOMMAND]
         foo destroy [SUBCOMMAND]
@@ -59,6 +60,7 @@ RSpec.describe "Rendering" do
     expected = <<~DESC
       Commands:
         foo assets [SUBCOMMAND]
+        foo callbacks DIR                      # Command with callbacks
         foo console                            # Starts Foo console
         foo db [SUBCOMMAND]
         foo destroy [SUBCOMMAND]
@@ -81,6 +83,7 @@ RSpec.describe "Rendering" do
     expected = <<~DESC
       Commands:
         foo assets [SUBCOMMAND]
+        foo callbacks DIR                      # Command with callbacks
         foo console                            # Starts Foo console
         foo db [SUBCOMMAND]
         foo destroy [SUBCOMMAND]