+++ spec/support/fixtures/registry.rb
@@ -1,0 +1,14 @@
+module Bar
+  module CLI
+    module Commands
+      extend Hanami::CLI::Registry
+
+      class Alpha < Hanami::CLI::Command
+        def call(*)
+        end
+      end
+
+      register "alpha", Alpha
+    end
+  end
+end