+++ spec/unit/hanami/cli/version_spec.rb
@@ -1,5 +1,5 @@
 RSpec.describe "Hanami::CLI::VERSION" do
   it "exposes version" do
-    expect(Hanami::CLI::VERSION).to eq("0.1.1")
+    expect(Hanami::CLI::VERSION).to eq("0.2.0")
   end
 end